Move old facade from agent to controller subpackage.
